refactor(eval): mv EvalStatus to structs/ & expose
Relocate EvalStatus struct and its ShowHydraStatus implementation from
queries/jobset.rs to structs/eval.rs. This centralizes struct definitions.
