+++ src/structs/mod.rs
@@ -4,6 +4,6 @@ mod icons;
 mod inputs;
 
 pub(crate) use build::BuildStatus;
-pub(crate) use eval::Evaluation;
+pub(crate) use eval::{EvalStatus, Evaluation};
 pub(crate) use icons::StatusIcon;
 pub(crate) use inputs::EvalInput;
